Add a parent key of puppetdb to gain parity with puppetserver metrics
Prior to this commit, the puppetdb metrics output placed a json
hash for each metric but they did not contain a parent key.

After this commit, all of the puppetdb metrics fall under a
puppetdb parent key so it is in the same form as the puppetserver
metrics.

In order to implement this, we add a puppetdb parent key and add a
comma after each possible metric but for the final output we
remove the last comma and output two closing braces.